Fix the client's available messages counter.
Adds a callback function to signal when a message is freed in the `MessageBus`.
This callback allows the client to correctly track request messages that have not been unref'd by `client.unref` but by somewhere else in the vsr machinery.

A subtle bug when a request is timed-out and its reply arrives when it is being retransmitted, causes the counter not to be incremented because the request is still referenced by the send queue.

The most common effect could be observed during `client.deinit()` when the assertion `self.messages_available == constants.client_request_queue_max` failed.